fastapi-app/auth.py
from passlib.context import CryptContext
from fastapi import Depends, HTTPException
from fastapi.security import OAuth2PasswordBearer
from jose import jwt
from jose.exceptions import ExpiredSignatureError, JWTError
import os
from datetime import datetime, timezone, timedelta

pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
SECRET_KEY = "SkillLink-secret"
oauth2_scheme = OAuth2PasswordBearer(tokenUrl="login")

def get_current_user(token: str = Depends(oauth2_scheme)):
    try:
        print("Received token: ", token)
        payload = jwt.decode(token, SECRET_KEY, algorithms=["HS256"])
        return payload
    except ExpiredSignatureError:
        raise HTTPException(status_code=401, detail="Token has expired.")
    except JWTError as e:
        print("Token decode error: ", e)
        raise HTTPException(status_code=401, detail="Invalid token.")

def hash_password(password: str):
    return pwd_context.hash(password)

def verify_password(plain_password, hashed_password):
    return pwd_context.verify(plain_password, hashed_password)

def create_access_token(data: dict, expires_delta: timedelta = timedelta(days=1)):
    to_encode = data.copy()
    expire = datetime.now(timezone.utc) + expires_delta
    to_encode.update({"exp": expire})
    return jwt.encode(to_encode, SECRET_KEY, algorithm="HS256")

def verify_org(current_user: dict = Depends(get_current_user)):
    if current_user["user_type"] != "org":
        raise HTTPException(status_code=403, detail="Only organizations can access this route.")
    return current_user

def verify_player(current_user: dict = Depends(get_current_user)):
    if current_user["user_type"] != "player":
        raise HTTPException(status_code=403, detail="Only players can access this route.")
    return current_user